Extract BrewBuildInspector and harden RHEL detection logic
