+++ doozer/doozerlib/assembly_inspector.py
@@ -10,7 +10,7 @@ from doozerlib.rpm_delivery import RPMDeliveries, RPMDelivery
 from doozerlib.rpm_utils import compare_nvr, parse_nvr
 
 from doozerlib import brew, util, Runtime
-from doozerlib.image import BrewBuildImageInspector
+from doozerlib.brew_info import BrewBuildImageInspector
 from doozerlib.rpmcfg import RPMMetadata
 from doozerlib.assembly import assembly_rhcos_config, AssemblyTypes, assembly_permits, AssemblyIssue, \
     AssemblyIssueCode, assembly_type
