+++ elliott/tests/test_metadata.py
@@ -37,6 +37,7 @@ class TestMetadata(unittest.TestCase):
         image_meta.get_component_name = Mock(return_value='foo-container')
         image_meta.branch_major_minor = Mock(return_value='4.7')
         image_meta.branch = Mock(return_value='rhaos-4.7-rhel-8')
+        image_meta.branch_el_target = Mock(return_value=8)
         image_meta.candidate_brew_tags = Mock(return_value=['rhaos-4.7-rhel-8-candidate', 'rhaos-4.7-rhel-7-candidate'])
 
         self.runtime = runtime
@@ -133,6 +134,32 @@ class TestMetadata(unittest.TestCase):
         builds = [
             self.build_record(now, assembly='not_ours'),
             self.build_record(now, assembly='stream')
+        ]
+        self.assertEqual(meta.get_latest_build(default=None), builds[1])
+
+        # If there is a build from the 'stream' assembly, but it is for the
+        # wrong RHEL based on our metadata tag, it should not be returned.
+        builds = [
+            self.build_record(now, assembly='not_ours', release_suffix='.el8'),
+            self.build_record(now, assembly='stream', release_suffix='.el9')
+        ]
+        self.assertIsNone(meta.get_latest_build(default=None))
+
+        # Filtering should prefer images which match our tag's RHEL version.
+        builds = [
+            self.build_record(now, assembly='not_ours'),
+            self.build_record(now, assembly='stream', release_suffix='.el8'),
+            self.build_record(now, assembly='stream', release_suffix='.el9')
+        ]
+        self.assertEqual(meta.get_latest_build(default=None), builds[1])
+
+        # Filtering should prefer images which match our tag's RHEL version, bug
+        # if there is no match for our elX, at the very least, it should filter out
+        # the wrong elY.
+        builds = [
+            self.build_record(now, assembly='not_ours'),
+            self.build_record(now, assembly='stream'),
+            self.build_record(now, assembly='stream', release_suffix='.el9')
         ]
         self.assertEqual(meta.get_latest_build(default=None), builds[1])
